Arreglando el inconveniente con los test, para validar bien las ejecuciones que fallan con su respectivo id de ejecucion
